fix: satisfy React Doctor for PierreDiff viewer
